# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Thor Firewall Control Plane API
  version: "0.3.0"
  description: |
    **Thor NGFW REST API** — الواجهة البرمجية الكاملة للتحكم في نظام Thor Firewall.

    ## Authentication
    All endpoints require authentication via:
    - `Bearer <JWT>` header
    - `X-API-Key: <key>` header

    ## Roles
    - `admin`    — Full access
    - `operator` — Read + write rules + blocklist
    - `viewer`   — Read only

    ## Rate Limits
    - General API: 100 req/min per IP
    - Query (AI):  10 req/min per IP
    - Auth:        5 req/min per IP
  contact:
    name: Thor Firewall Team
    url:  https://github.com/mhmsdfhwhegggggggg/Thor-Firewall
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Versioned API
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /analytics/top-threats:
    get:
      summary: Top threat sources in time window
      tags: [Analytics]
      parameters:
        - name: hours
          in: query
          schema: { type: integer, default: 24 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    src_ip:     { type: string }
                    count:      { type: integer }
                    max_risk:   { type: number }
                    threat_type: { type: string }
                    geo_country: { type: string }

  # ── AI Query ──────────────────────────────────────────────────────────────
...